Add mark at output with change
To secure that fee is charged from spending coins and not from change,
DeStream marks output with change.
It is implemented via additional empty input with PrevOut hash
uint256.Zero, that points to output with change. Input is signed, so
it's verified that user has access to address from output, and this is
change address.